refactor(LibInputBackend): Move HandleTouch in `LibInputBackend.Tocuch.cs`
