+++ src/Linux/Avalonia.LinuxFramebuffer/Input/LibInput/LibInputBackend.cs
@@ -11,10 +11,8 @@ namespace Avalonia.LinuxFramebuffer.Input.LibInput
     {
         private IScreenInfoProvider _screen;
         private IInputRoot _inputRoot;
-        private TouchDevice _touch = new TouchDevice();
         private const string LibInput = nameof(Avalonia.LinuxFramebuffer) + "/" + nameof(Avalonia.LinuxFramebuffer.Input) + "/" + nameof(LibInput);
         private Action<RawInputEventArgs> _onInput;
-        private Dictionary<int, Point> _pointers = new Dictionary<int, Point>();
 
         public LibInputBackend()
         {
@@ -57,44 +55,6 @@ namespace Avalonia.LinuxFramebuffer.Input.LibInput
         }
 
         private void ScheduleInput(RawInputEventArgs ev) => _onInput.Invoke(ev);
-
-        private void HandleTouch(IntPtr ev, LibInputEventType type)
-        {
-            var tev = libinput_event_get_touch_event(ev);
-            if (tev == IntPtr.Zero)
-                return;
-            if (type < LibInputEventType.LIBINPUT_EVENT_TOUCH_FRAME)
-            {
-                var info = _screen.ScaledSize;
-                var slot = libinput_event_touch_get_slot(tev);
-                Point pt;
-
-                if (type == LibInputEventType.LIBINPUT_EVENT_TOUCH_DOWN
-                    || type == LibInputEventType.LIBINPUT_EVENT_TOUCH_MOTION)
-                {
-                    var x = libinput_event_touch_get_x_transformed(tev, (int)info.Width);
-                    var y = libinput_event_touch_get_y_transformed(tev, (int)info.Height);
-                    pt = new Point(x, y);
-                    _pointers[slot] = pt;
-                }
-                else
-                {
-                    _pointers.TryGetValue(slot, out pt);
-                    _pointers.Remove(slot);
-                }
-
-                var ts = libinput_event_touch_get_time_usec(tev) / 1000;
-                if (_inputRoot == null)
-                    return;
-                ScheduleInput(new RawTouchEventArgs(_touch, ts,
-                    _inputRoot,
-                    type == LibInputEventType.LIBINPUT_EVENT_TOUCH_DOWN ? RawPointerEventType.TouchBegin
-                    : type == LibInputEventType.LIBINPUT_EVENT_TOUCH_UP ? RawPointerEventType.TouchEnd
-                    : type == LibInputEventType.LIBINPUT_EVENT_TOUCH_MOTION ? RawPointerEventType.TouchUpdate
-                    : RawPointerEventType.TouchCancel,
-                    pt, RawInputModifiers.None, slot));
-            }
-        }
 
         public void Initialize(IScreenInfoProvider screen, Action<RawInputEventArgs> onInput)
         {
